Fix errors when building on `ROS2 Humble`
Missing declerations for the dependency `rclcpp_components` prevented
some headers from being found and caused errors when building on `ROS2
Humble Hawksbill`.

The header paths have changed in `ROS2 Humble`. This change and the
reasons behind it have been documented at:
https://docs.ros.org/en/rolling/Releases/Release-Humble-Hawksbill.html#c-headers-are-installed-in-a-subdirectory
Although it's not certain that the issues were related to this change,
they were not observed on the `ROS2 Galactic`.

CREDIT:
  The same changes were first made to the original repository in the commit
https://github.com/clydemcqueen/tello_ros/commit/c2b7b47d681513f0fc368d68b700910536d7b97c
to which I stumbled upon while looking for a fix to the issues. Credit
for the fix belongs thus to the original author.